fix: 32-bit-safe wide-field extraction across the CNAV-family parsers
Sweep for the same defect classes the BeiDou review turned up:
- get_twos_complement_num sign-extended through the platform UInt, which
  overflows for the >32-bit fields (the 33-bit GPS CNAV/CNAV-2 and BeiDou
  B-CNAV angle fields) on 32-bit Julia; sign-extend through UInt64.
- The GPS CNAV and CNAV-2 eccentricity parsers narrowed their 33-bit field
  through a platform Int, exactly the defect already fixed in the BeiDou
  B1C parser; convert through Int64 like the B-CNAV parsers.
- Drop stray inline-comment text that had leaked into two B2b constant
  docstrings.

The Chinese-vs-English cross-check that exposed the B2b SOW misprint was
also run against the B1I/B3I, B1C, and B2a ICDs (parameter-table bit
counts, two's-complement markers, scale factors): no further
discrepancies — the B2b SOW row remains the only translation error
affecting decoding.

+++ src/gps/cnav.jl
@@ -1133,7 +1133,9 @@ function parse_mt10(raw::GPSCNAVData, word::UInt320, PI::Float64)
         Δn_0 = get_twos_complement_num(word, word_length, 133, 17) * 2.0^-44 * PI,
         Δn_0_dot = get_twos_complement_num(word, word_length, 150, 23) * 2.0^-57 * PI,
         M_0 = get_twos_complement_num(word, word_length, 173, 33) * 2.0^-32 * PI,
-        e = Int(get_bits(word, word_length, 206, 33)) * 2.0^-34,
+        # 33-bit field: convert through Int64 explicitly (a platform Int
+        # would overflow on 32-bit), matching the BeiDou B-CNAV parsers.
+        e = Int64(get_bits(word, word_length, 206, 33)) * 2.0^-34,
         ω = get_twos_complement_num(word, word_length, 239, 33) * 2.0^-32 * PI,
         integrity_status_flag = get_bit(word, word_length, 272),
         l2c_phasing = get_bit(word, word_length, 273),
